Implement remak plugin to replace version string in docs
We provide documentation for few versions of
Inspektor Gadget. Before this commit we're always
using "latest" in all the documentation, it was a
bit misleading as the features of a component can
change from one version to the other.

This commit adds a remark plugin that replaces
some constants with the actual version of the
documentation.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>